Move data loading function to new source file.
